// src/interfaces/seriesInterface.ts
// Generated by https://quicktype.io

export interface Series {
  id: number;
  title: string;
  description: null;
  resourceURI: string;
  urls: URL[];
  startYear: number;
  endYear: number;
  rating: string;
  type: string;
  modified: string;
  thumbnail: Thumbnail;
  creators: Characters;
  characters: Characters;
  stories: Characters;
  comics: Characters;
  events: Characters;
  next: null;
  previous: null;
}

export interface Characters {
  available: number;
  collectionURI: string;
  items: Item[];
  returned: number;
}

export interface Item {
  resourceURI: string;
  name: string;
  role?: string;
  type?: string;
}

export interface Thumbnail {
  path: string;
  extension: string;
}

export interface URL {
  type: string;
  url: string;
}